hit_tux takes a float for damage, not an integer. This fixes overheating doing no damage.
git-svn-id: https://freedroid.svn.sourceforge.net/svnroot/freedroid@3430 524c2bfb-fa3f-0410-b5a2-92f4e1dc5b61

+++ src/influ.c
@@ -748,7 +748,7 @@ void move_tux()
  * This function decrements Tux's health and increments the relevent statistic 
  * variable.
  */
-void hit_tux(int damage, int owner)
+void hit_tux(float damage, int owner)
 {
 	Me.energy -= damage;
